Time: 101 ms (47.71%) | Memory: 8.7 MB (59.58%) - LeetSync

+++ 1-two-sum/two-sum.c
@@ -3,8 +3,8 @@
  */
 int* twoSum(int* nums, int numsSize, int target, int* returnSize) {
     int *res=malloc(50);
-    for(int outer=0;outer<numsSize;outer++){
-        for(int inner=outer+1;inner<numsSize;inner++){
+    for(long outer=0;outer<numsSize;outer++){
+        for(long inner=outer+1;inner<numsSize;inner++){
        if(nums[outer]+nums[inner]==target){
         res[0]=outer;
         res[1]=inner;
